requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
cssselect==1.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1.13
django-admin-interface==0.28.1
django-ckeditor==6.7.0
django-colorfield==0.11.0
django-cors-headers==4.2.0
django-countries==7.5.1
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-import-export==3.2.0
django-js-asset==2.1.0
django-prometheus==2.3.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.4.2
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.2.1
et-xmlfile==1.1.0
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.9.1
geventhttpclient==2.0.11
graphviz==0.20.1
greenlet==3.0.3
idna==3.6
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.4
locust==2.20.1
lxml==4.9.3
Markdown==3.4.3
markdown2==2.4.10
MarkupPy==1.14
MarkupSafe==2.1.3
msgpack==1.0.7
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.0.0
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2-binary==2.9.9
pydot==1.4.2
PyJWT==2.7.0
pymongo==4.6.0
pyparsing==3.1.0
pyquery==2.0.0
pytest==7.4.4
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.2
readtime==3.0.0
redis==5.0.1
referencing==0.33.0
requests==2.31.0
roundrobin==0.0.4
rpds-py==0.17.1
shortuuid==1.0.11
six==1.16.0
soupsieve==2.5
sqlparse==0.2.4
tablib==3.5.0
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.16.0
zope.event==5.0
zope.interface==6.1
gunicorn==23.0.0
django-admin-interface==0.28.1
